requirements.txt
streamlit>=1.28.0
rembg>=2.0.50
pillow>=10.0.0
numpy>=1.24.0
